Estimate parameters and their covariance from a set of estimating equations

Estimate parameters and their covariance from a set of estimating equations

estimate_equations(eeFUN, data, units, corrections_list = NULL,
  numDeriv_options = list(method = "Richardson"),
  rootsolver = rootSolve::multiroot, rootsolver_options = NULL,
  findroots = TRUE, roots = NULL, ee_args = NULL, compute_vcov = TRUE,
  ...)

Arguments

eeFUN
a function that takes in group-level data and returns a function that takes parameters as its first argument

data
a data.frame

units
a string identifying the grouping variable in data

numDeriv_options
a list of options for jacobian

rootsolver
the function that will find roots of the estimating equations when findroots = TRUE.

rootsolver_options
a list of options for the rootsolver function

findroots
whether or not to find the roots of the estimating equations. Defaults to TRUE.

roots
a numeric vector containing either starting values for the roots when using the default rootsolver or roots that have been estimated elsewhere

...
additional arguments passed to the eeFUN.

corrections
character vector of small sample corrections to apply

correction_options
a list of options for the small sample corrections

Value

a list with the following

parameters - a vector of estimated parameters

vcov - the variance-covariance matrix for the parameters