Creates an m_estimation_basis object

Creates an m_estimation_basis object

create_basis(estFUN, data, units, outer_args, inner_args)

Arguments

estFUN
a function that takes in group-level data and returns a function that takes parameters as its first argument

data
a data.frame

units
an optional character string identifying the grouping variable in data

outer_args
a list of arguments passed to the outer (data) function of estFUN. (optional)

inner_args
a list of arguments passed to the inner (theta) function of estFUN. (optional)

Value

a m_estimation_basis

Details

Either data or split_data must be provided

Examples

myee <- function(data){
   function(theta){
    c(data$Y1 - theta[1],
     (data$Y1 - theta[1])^2 - theta[2])
   }
 }
mybasis <- create_basis(
   estFUN = myee,
   data   = geexex)
mybasis

#> An object of class m_estimation_basis
#> psi: 
#> {
#>     function(theta) {
#>         c(data$Y1 - theta[1], (data$Y1 - theta[1])^2 - theta[2])
#>     }
#> }
#> Data:
#>           Y1        Y2       X1       Y3       W1        Z1 X2         Y4 Y5
#> 1  3.6683066 2.0281718 4.949316 16.34576 4.823768  8.921782  0 0.09273926  1
#> 2 10.4524548 1.6432966 7.851962 25.68742 7.884845 13.909474  0 1.01672736  1
#> 3  3.1234106 2.8526264 4.729075 16.10831 4.709346  9.014695  0 0.49399039  0
#> 4  8.3715025 2.5133652 2.564395 10.57997 2.786091  6.733378  0 1.24322433  0
#> 5 -0.8319749 3.0182030 4.782347 16.46401 4.811590  9.290492  0 0.69520599  1
#> 6  3.3987763 0.9785209 5.335713 18.32577 5.415370 10.322199  0 0.95220138  1
#> Units: