Estimate parameters and their covariance from a set of estimating equations

Estimate parameters and their covariance from a set of estimating equations

estimate_equations(eeFUN, data, units = NULL, weights = NULL,
  roots = NULL, outer_eeargs = NULL, inner_eeargs = NULL,
  compute_roots = TRUE, compute_vcov = TRUE, corrections_list = NULL,
  derivFUN = numDeriv::jacobian, derivFUN_control = list(method =
  "Richardson"), rootFUN = rootSolve::multiroot, rootFUN_control = NULL,
  rootFUN_object = "root", approxFUN = NULL, approxFUN_control = NULL)

Arguments

eeFUN
a function that takes in group-level data and returns a function that takes parameters as its first argument

data
a data.frame

units
a string identifying the grouping variable in data

roots
a numeric vector containing either starting values for the roots when using the default rootsolver or roots that have been estimated elsewhere

outer_eeargs
a list of arguments passed to the outer (data) function of eeFUN. (optional)

inner_eeargs
a list of arguments passed to the inner (theta) function of eeFUN. (optional)

compute_roots
whether or not to find the roots of the estimating equations. Defaults to TRUE.

compute_vcov
whether or not to compute the variance-covariance matrix. Defaults to TRUE.

corrections_list
an optional list of small sample corrections where each list element is a list with two elements: `fun` and `options`. See details.

derivFUN
the function used to take derivatives of the estimating equation functions. Defaults to numDeriv::jacobian

derivFUN_control
a list of options passed to jacobian (or the derivFUN function).

rootFUN
the function used to find roots of the estimating equations. Defaults to multiroot.

rootFUN_control
a list of options to be passed to the rootsolver function

rootFUN_object
the name of the object within the output of rootFUN that contains the parameters estimates. For example, the 'root' object within the output of multiroot::rootSolve contains the parameter estimates. Defaults to 'root'. Can also be a set of numeric positions within the object.

approxFUN
a function that approximates the inner function of eeFUN

approxFUN_control
arguments passed to approxFUN

Value

a list with the following

parameters - a vector of estimated parameters

vcov - the variance-covariance matrix for the parameters

corrections - a list of corrected variance-covariance matrices

eeFUN arguments

Additional arguments may be passed to both the inner and outer function of the `eeFUN`. Elements in an outer_eeargs listare passed to the outer function; any elements of the inner_eeargs list are passed to the inner function. For example, a practical example might be computing a counterfactual mean using an IPW estimator:

myeeFUN <- function(data, model){
  X <- model.matrix(model, data = data) #covariates
  A <- data$A #treatment
  Y <- data$Y #outcome
  p <- ncol(X) #number of parameters in model
  function(theta, a){
    Y * (A == a) * 1/plogis(X %*% theta[p - 1]) - theta[p]
    # Here theta[p] is the target parameter.
  }
}

Then to estimate the mean where `a == 1`:

estimate_equations(
  eeFUN = myeeFUN,
  data  = mydata,
  units = myunits,
  inner_eeargs = list(a = 1),
  outer_eeargs = list(model = mymodel)
)