Estimate component matrices of the empirical sandwich covariance estimator

For a given set of estimating equations computes the 'meat' (\(B_m\) in Stefanski and Boos notation) and 'bread' (\(A_m\) in Stefanski and Boos notation) matrices necessary to compute the covariance matrix.

estimate_sandwich_matrices(.basis, .theta)

Arguments

.basis
basis an object of class m_estimation_basis

.theta
vector of parameter estimates (i.e. estimated roots)

Value

a sandwich_components object

Details

For a set of estimating equations (\(\sum_i \psi(O_i, \theta) = 0\)), this function computes:

$$A_i = \partial \psi(O_i, \theta)/\partial \theta$$

$$A = \sum_i A_i$$

$$B_i = \psi(O_i, \theta)\psi(O_i, \theta)^T$$

$$B = \sum_i B_i$$

where all of the above are evaluated at \(\hat{\theta}\). The partial derivatives in \(A_i\) numerically approximated by the function defined in deriv_control.

Note that \(A = \sum_i A_i\) and not \(\sum_i A_i/m\), and the same for \(B\).

References

Stefanski, L. A., & Boos, D. D. (2002). The calculus of m-estimation. The American Statistician, 56(1), 29-38.

Examples

myee <- function(data){
  function(theta){
    c(data$Y1 - theta[1],
     (data$Y1 - theta[1])^2 - theta[2])
   }
 }

# Start with a basic basis
mybasis <- create_basis(
  estFUN = myee,
  data   = geexex)

# Now estimate sandwich matrices
estimate_sandwich_matrices(
 mybasis, c(mean(geexex$Y1), var(geexex$Y1)))

#> An object of class "sandwich_components"
#> Slot ".A":
#>               [,1] [,2]
#> [1,]  1.000000e+02    0
#> [2,] -1.586615e-10  100
#> 
#> Slot ".A_i":
#> $`1`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.752514    1
#> 
#> $`2`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] 10.81578    1
#> 
#> $`3`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.842305    1
#> 
#> $`4`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 6.653878    1
#> 
#> $`5`
#>           [,1] [,2]
#> [1,]   1.00000    0
#> [2,] -11.75308    1
#> 
#> $`6`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.291574    1
#> 
#> $`7`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -6.300465    1
#> 
#> $`8`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.043499    1
#> 
#> $`9`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 9.831685    1
#> 
#> $`10`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.9485972    1
#> 
#> $`11`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 1.291621    1
#> 
#> $`12`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 1.947683    1
#> 
#> $`13`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -5.005397    1
#> 
#> $`14`
#>           [,1] [,2]
#> [1,]   1.00000    0
#> [2,] -11.52285    1
#> 
#> $`15`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] -2.73693    1
#> 
#> $`16`
#>           [,1] [,2]
#> [1,] 1.0000000    0
#> [2,] 0.9379618    1
#> 
#> $`17`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 8.055833    1
#> 
#> $`18`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.133716    1
#> 
#> $`19`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.509455    1
#> 
#> $`20`
#>         [,1] [,2]
#> [1,]  1.0000    0
#> [2,] 12.8324    1
#> 
#> $`21`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -6.278834    1
#> 
#> $`22`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 3.302893    1
#> 
#> $`23`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -4.760703    1
#> 
#> $`24`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 3.231159    1
#> 
#> $`25`
#>           [,1] [,2]
#> [1,]   1.00000    0
#> [2,] -12.45122    1
#> 
#> $`26`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -4.239123    1
#> 
#> $`27`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.327459    1
#> 
#> $`28`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 7.970532    1
#> 
#> $`29`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.845686    1
#> 
#> $`30`
#>         [,1] [,2]
#> [1,] 1.00000    0
#> [2,] 2.29405    1
#> 
#> $`31`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.431482    1
#> 
#> $`32`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -6.892173    1
#> 
#> $`33`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 5.423243    1
#> 
#> $`34`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.770696    1
#> 
#> $`35`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] 10.69635    1
#> 
#> $`36`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 3.455444    1
#> 
#> $`37`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -1.296536    1
#> 
#> $`38`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 3.555264    1
#> 
#> $`39`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.4103641    1
#> 
#> $`40`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 3.559842    1
#> 
#> $`41`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.356527    1
#> 
#> $`42`
#>           [,1] [,2]
#> [1,]   1.00000    0
#> [2,] -17.18108    1
#> 
#> $`43`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 1.165449    1
#> 
#> $`44`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 5.191265    1
#> 
#> $`45`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -7.943802    1
#> 
#> $`46`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -8.998276    1
#> 
#> $`47`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.587908    1
#> 
#> $`48`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -4.218017    1
#> 
#> $`49`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 3.262841    1
#> 
#> $`50`
#>           [,1] [,2]
#> [1,] 1.0000000    0
#> [2,] 0.9841168    1
#> 
#> $`51`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 8.188365    1
#> 
#> $`52`
#>         [,1] [,2]
#> [1,]  1.0000    0
#> [2,] 13.1389    1
#> 
#> $`53`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.2327012    1
#> 
#> $`54`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.2827533    1
#> 
#> $`55`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 1.912849    1
#> 
#> $`56`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.786123    1
#> 
#> $`57`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.9959563    1
#> 
#> $`58`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] -0.88019    1
#> 
#> $`59`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 2.023568    1
#> 
#> $`60`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 1.022743    1
#> 
#> $`61`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.027283    1
#> 
#> $`62`
#>           [,1] [,2]
#> [1,] 1.0000000    0
#> [2,] 0.3831244    1
#> 
#> $`63`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -1.507302    1
#> 
#> $`64`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 6.268315    1
#> 
#> $`65`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.0352244    1
#> 
#> $`66`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -5.127449    1
#> 
#> $`67`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.109045    1
#> 
#> $`68`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -5.612504    1
#> 
#> $`69`
#>         [,1] [,2]
#> [1,] 1.00000    0
#> [2,] 1.53121    1
#> 
#> $`70`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 6.681925    1
#> 
#> $`71`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 4.967466    1
#> 
#> $`72`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 1.522182    1
#> 
#> $`73`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.8176918    1
#> 
#> $`74`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 2.226746    1
#> 
#> $`75`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.5266062    1
#> 
#> $`76`
#>           [,1] [,2]
#> [1,]   1.00000    0
#> [2,] -16.42393    1
#> 
#> $`77`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 2.777827    1
#> 
#> $`78`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.069934    1
#> 
#> $`79`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] 10.05734    1
#> 
#> $`80`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] 17.25968    1
#> 
#> $`81`
#>           [,1] [,2]
#> [1,]   1.00000    0
#> [2,] -10.00691    1
#> 
#> $`82`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 4.629864    1
#> 
#> $`83`
#>           [,1] [,2]
#> [1,] 1.0000000    0
#> [2,] 0.9030276    1
#> 
#> $`84`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -4.278789    1
#> 
#> $`85`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 4.872697    1
#> 
#> $`86`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 5.576646    1
#> 
#> $`87`
#>            [,1] [,2]
#> [1,]  1.0000000    0
#> [2,] -0.8347135    1
#> 
#> $`88`
#>         [,1] [,2]
#> [1,]  1.0000    0
#> [2,] -2.4507    1
#> 
#> $`89`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -8.775649    1
#> 
#> $`90`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -5.087647    1
#> 
#> $`91`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -1.953179    1
#> 
#> $`92`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.095662    1
#> 
#> $`93`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 7.542036    1
#> 
#> $`94`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -2.216907    1
#> 
#> $`95`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] 15.07308    1
#> 
#> $`96`
#>           [,1] [,2]
#> [1,]  1.000000    0
#> [2,] -3.529053    1
#> 
#> $`97`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] 12.51525    1
#> 
#> $`98`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 1.206403    1
#> 
#> $`99`
#>          [,1] [,2]
#> [1,]  1.00000    0
#> [2,] -8.77275    1
#> 
#> $`100`
#>          [,1] [,2]
#> [1,] 1.000000    0
#> [2,] 4.526372    1
#> 
#> 
#> Slot ".B":
#>           [,1]       [,2]
#> [1,] 1004.1239   366.7969
#> [2,]  366.7969 24922.9925
#> 
#> Slot ".B_i":
#> $`1`
#>           [,1]     [,2]
#> [1,]  1.894083 11.35217
#> [2,] 11.352168 68.03912
#> 
#> $`2`
#>           [,1]     [,2]
#> [1,]  29.24529 103.3049
#> [2,] 103.30492 364.9103
#> 
#> $`3`
#>           [,1]     [,2]
#> [1,]  3.690828 12.39497
#> [2,] 12.394966 41.62621
#> 
#> $`4`
#>           [,1]      [,2]
#> [1,] 11.068524 3.0802759
#> [2,]  3.080276 0.8572145
#> 
#> $`5`
#>           [,1]      [,2]
#> [1,]   34.5337 -143.3349
#> [2,] -143.3349  594.9226
#> 
#> $`6`
#>           [,1]     [,2]
#> [1,]  2.708615 12.23486
#> [2,] 12.234864 55.26511
#> 
#> $`7`
#>           [,1]       [,2]
#> [1,] 9.9239647 0.68895834
#> [2,] 0.6889583 0.04783004
#> 
#> $`8`
#>           [,1]     [,2]
#> [1,]  2.315721 11.91065
#> [2,] 11.910648 61.26106
#> 
#> $`9`
#>          [,1]      [,2]
#> [1,] 24.16551  68.93408
#> [2,] 68.93408 196.64009
#> 
#> $`10`
#>           [,1]      [,2]
#> [1,] 0.2249591  4.703954
#> [2,] 4.7039541 98.360900
#> 
#> $`11`
#>            [,1]      [,2]
#> [1,]  0.4170714 -6.280893
#> [2,] -6.2808925 94.587180
#> 
#> $`12`
#>            [,1]      [,2]
#> [1,]  0.9483677 -8.953791
#> [2,] -8.9537909 84.535112
#> 
#> $`13`
#>          [,1]      [,2]
#> [1,] 6.263501  9.708381
#> [2,] 9.708381 15.047921
#> 
#> $`14`
#>           [,1]      [,2]
#> [1,]   33.1940 -132.8085
#> [2,] -132.8085  531.3642
#> 
#> $`15`
#>           [,1]     [,2]
#> [1,]  1.872697 11.31716
#> [2,] 11.317164 68.39238
#> 
#> $`16`
#>            [,1]      [,2]
#> [1,]  0.2199431 -4.653567
#> [2,] -4.6535674 98.460420
#> 
#> $`17`
#>          [,1]     [,2]
#> [1,] 16.22411 24.49556
#> [2,] 24.49556 36.98400
#> 
#> $`18`
#>          [,1]      [,2]
#> [1,] 1.138186  9.606502
#> [2,] 9.606502 81.080647
#> 
#> $`19`
#>           [,1]     [,2]
#> [1,]  1.574341 10.75091
#> [2,] 10.750912 73.41619
#> 
#> $`20`
#>           [,1]     [,2]
#> [1,]  41.16761 199.0623
#> [2,] 199.06226 962.5475
#> 
#> $`21`
#>           [,1]       [,2]
#> [1,] 9.8559376 0.90015836
#> [2,] 0.9001584 0.08221289
#> 
#> $`22`
#>            [,1]      [,2]
#> [1,]   2.727275 -12.24612
#> [2,] -12.246119  54.98802
#> 
#> $`23`
#>           [,1]     [,2]
#> [1,]  5.666072 10.65586
#> [2,] 10.655865 20.03989
#> 
#> $`24`
#>            [,1]      [,2]
#> [1,]   2.610097 -12.16946
#> [2,] -12.169463  56.73959
#> 
#> $`25`
#>            [,1]      [,2]
#> [1,]   38.75822 -178.1493
#> [2,] -178.14928  818.8499
#> 
#> $`26`
#>          [,1]     [,2]
#> [1,]  4.49254 11.97579
#> [2,] 11.97579 31.92391
#> 
#> $`27`
#>           [,1]     [,2]
#> [1,]  1.354266 10.22732
#> [2,] 10.227320 77.23596
#> 
#> $`28`
#>          [,1]     [,2]
#> [1,] 15.88235 22.87416
#> [2,] 22.87416 32.94394
#> 
#> $`29`
#>           [,1]     [,2]
#> [1,]  3.697326 12.39338
#> [2,] 12.393377 41.54240
#> 
#> $`30`
#>            [,1]      [,2]
#> [1,]   1.315666 -10.12479
#> [2,] -10.124788  77.91592
#> 
#> $`31`
#>           [,1]     [,2]
#> [1,]  2.943767 12.35145
#> [2,] 12.351445 51.82413
#> 
#> $`32`
#>           [,1]      [,2]
#> [1,] 11.875512 -5.971538
#> [2,] -5.971538  3.002756
#> 
#> $`33`
#>           [,1]      [,2]
#> [1,]  7.352891 -7.564813
#> [2,] -7.564813  7.782843
#> 
#> $`34`
#>           [,1]     [,2]
#> [1,]  3.554538 12.42091
#> [2,] 12.420915 43.40343
#> 
#> $`35`
#>          [,1]      [,2]
#> [1,] 28.60297  98.72891
#> [2,] 98.72891 340.78273
#> 
#> $`36`
#>            [,1]      [,2]
#> [1,]   2.985024 -12.36642
#> [2,] -12.366416  51.23183
#> 
#> $`37`
#>           [,1]      [,2]
#> [1,] 0.4202515  6.302731
#> [2,] 6.3027308 94.525334
#> 
#> $`38`
#>            [,1]      [,2]
#> [1,]   3.159976 -12.41265
#> [2,] -12.412653  48.75796
#> 
#> $`39`
#>            [,1]       [,2]
#> [1,] 0.04209968   2.072455
#> [2,] 2.07245503 102.021430
#> 
#> $`40`
#>            [,1]      [,2]
#> [1,]   3.168118 -12.41414
#> [2,] -12.414142  48.64431
#> 
#> $`41`
#>           [,1]     [,2]
#> [1,]  2.816568 12.29512
#> [2,] 12.295121 53.67170
#> 
#> $`42`
#>            [,1]      [,2]
#> [1,]   73.79736 -546.8281
#> [2,] -546.82809 4051.9197
#> 
#> $`43`
#>            [,1]      [,2]
#> [1,]  0.3395676 -5.712504
#> [2,] -5.7125035 96.100728
#> 
#> $`44`
#>           [,1]      [,2]
#> [1,]  6.737307 -8.839059
#> [2,] -8.839059 11.596468
#> 
#> $`45`
#>           [,1]      [,2]
#> [1,]  15.77600 -22.37504
#> [2,] -22.37504  31.73443
#> 
#> $`46`
#>           [,1]     [,2]
#> [1,]  20.24224 -45.4394
#> [2,] -45.43940 102.0015
#> 
#> $`47`
#>           [,1]     [,2]
#> [1,]  3.218271 12.42205
#> [2,] 12.422045 47.94724
#> 
#> $`48`
#>           [,1]     [,2]
#> [1,]  4.447916 12.01027
#> [2,] 12.010274 32.43017
#> 
#> $`49`
#>            [,1]      [,2]
#> [1,]   2.661533 -12.20487
#> [2,] -12.204873  55.96734
#> 
#> $`50`
#>            [,1]      [,2]
#> [1,]  0.2421215 -4.871646
#> [2,] -4.8716459 98.020772
#> 
#> $`51`
#>          [,1]     [,2]
#> [1,] 16.76233 27.10211
#> [2,] 27.10211 43.81996
#> 
#> $`52`
#>           [,1]      [,2]
#> [1,]  43.15767  216.8904
#> [2,] 216.89039 1089.9904
#> 
#> $`53`
#>            [,1]      [,2]
#> [1,] 0.01353747   1.17853
#> [2,] 1.17853029 102.59924
#> 
#> $`54`
#>            [,1]      [,2]
#> [1,] 0.01998735   1.43111
#> [2,] 1.43111012 102.46861
#> 
#> $`55`
#>            [,1]      [,2]
#> [1,]  0.9147474 -8.825805
#> [2,] -8.8258048 85.154474
#> 
#> $`56`
#>          [,1]     [,2]
#> [1,]  1.94062 11.42595
#> [2,] 11.42595 67.27355
#> 
#> $`57`
#>           [,1]      [,2]
#> [1,] 0.2479823  4.927336
#> [2,] 4.9273362 97.904757
#> 
#> $`58`
#>           [,1]      [,2]
#> [1,] 0.1936836  4.378497
#> [2,] 4.3784973 98.982241
#> 
#> $`59`
#>           [,1]      [,2]
#> [1,]  1.023707 -9.226416
#> [2,] -9.226416 83.155409
#> 
#> $`60`
#>            [,1]      [,2]
#> [1,]  0.2615007 -5.052945
#> [2,] -5.0529451 97.637418
#> 
#> $`61`
#>          [,1]     [,2]
#> [1,] 1.027469  9.23954
#> [2,] 9.239540 83.08681
#> 
#> $`62`
#>             [,1]       [,2]
#> [1,]  0.03669607  -1.935922
#> [2,] -1.93592154 102.130619
#> 
#> $`63`
#>           [,1]      [,2]
#> [1,] 0.5679896  7.215962
#> [2,] 7.2159624 91.674420
#> 
#> $`64`
#>           [,1]       [,2]
#> [1,]  9.822945 -1.0020556
#> [2,] -1.002056  0.1022214
#> 
#> $`65`
#>              [,1]        [,2]
#> [1,] 0.0003101896   0.1786292
#> [2,] 0.1786291886 102.8673716
#> 
#> $`66`
#>          [,1]      [,2]
#> [1,] 6.572683  9.152451
#> [2,] 9.152451 12.744772
#> 
#> $`67`
#>          [,1]      [,2]
#> [1,] 1.112018  9.523021
#> [2,] 9.523021 81.552601
#> 
#> $`68`
#>         [,1]     [,2]
#> [1,] 7.87505 6.363500
#> [2,] 6.36350 5.142079
#> 
#> $`69`
#>            [,1]      [,2]
#> [1,]  0.5861514 -7.316517
#> [2,] -7.3165173 91.326963
#> 
#> $`70`
#>           [,1]     [,2]
#> [1,] 11.162030 3.405658
#> [2,]  3.405658 1.039104
#> 
#> $`71`
#>           [,1]     [,2]
#> [1,]  6.168929 -9.86970
#> [2,] -9.869700 15.79058
#> 
#> $`72`
#>            [,1]     [,2]
#> [1,]  0.5792591 -7.27862
#> [2,] -7.2786202 91.45874
#> 
#> $`73`
#>          [,1]      [,2]
#> [1,] 0.167155  4.078447
#> [2,] 4.078447 99.510810
#> 
#> $`74`
#>           [,1]      [,2]
#> [1,]  1.239600 -9.912434
#> [2,] -9.912434 79.264580
#> 
#> $`75`
#>            [,1]       [,2]
#> [1,] 0.06932852   2.652341
#> [2,] 2.65234089 101.472118
#> 
#> $`76`
#>            [,1]      [,2]
#> [1,]   67.43633 -470.4935
#> [2,] -470.49346 3282.5643
#> 
#> $`77`
#>            [,1]      [,2]
#> [1,]   1.929081 -11.40796
#> [2,] -11.407959  67.46297
#> 
#> $`78`
#>           [,1]     [,2]
#> [1,]  2.356123 11.95208
#> [2,] 11.952084 60.63025
#> 
#> $`79`
#>          [,1]      [,2]
#> [1,] 25.28754  76.15862
#> [2,] 76.15862 229.36729
#> 
#> $`80`
#>          [,1]      [,2]
#> [1,]  74.4741  555.1699
#> [2,] 555.1699 4138.5338
#> 
#> $`81`
#>           [,1]      [,2]
#> [1,]  25.03456 -74.51092
#> [2,] -74.51092 221.76849
#> 
#> $`82`
#>            [,1]      [,2]
#> [1,]   5.358911 -11.07407
#> [2,] -11.074067  22.88431
#> 
#> $`83`
#>            [,1]      [,2]
#> [1,]  0.2038647 -4.487506
#> [2,] -4.4875058 98.779761
#> 
#> $`84`
#>           [,1]     [,2]
#> [1,]  4.577009 11.90714
#> [2,] 11.907135 30.97653
#> 
#> $`85`
#>            [,1]      [,2]
#> [1,]   5.935795 -10.24940
#> [2,] -10.249404  17.69776
#> 
#> $`86`
#>           [,1]      [,2]
#> [1,]  7.774745 -6.602527
#> [2,] -6.602527  5.607047
#> 
#> $`87`
#>           [,1]      [,2]
#> [1,] 0.1741867  4.160412
#> [2,] 4.1604120 99.370571
#> 
#> $`88`
#>           [,1]     [,2]
#> [1,]  1.501483 10.58847
#> [2,] 10.588474 74.67004
#> 
#> $`89`
#>           [,1]      [,2]
#> [1,]  19.25300 -39.97455
#> [2,] -39.97455  82.99823
#> 
#> $`90`
#>          [,1]      [,2]
#> [1,] 6.471038  9.339972
#> [2,] 9.339972 13.480846
#> 
#> $`91`
#>           [,1]      [,2]
#> [1,] 0.9537271  8.973821
#> [2,] 8.9738210 84.436589
#> 
#> $`92`
#>          [,1]      [,2]
#> [1,] 1.097949  9.477332
#> [2,] 9.477332 81.806890
#> 
#> $`93`
#>          [,1]     [,2]
#> [1,] 14.22058 15.37788
#> [2,] 15.37788 16.62936
#> 
#> $`94`
#>          [,1]     [,2]
#> [1,] 1.228669  9.88075
#> [2,] 9.880750 79.45934
#> 
#> $`95`
#>           [,1]      [,2]
#> [1,]  56.79944  351.6307
#> [2,] 351.63068 2176.8548
#> 
#> $`96`
#>           [,1]     [,2]
#> [1,]  3.113554 12.40305
#> [2,] 12.403054 49.40840
#> 
#> $`97`
#>           [,1]     [,2]
#> [1,]  39.15787 181.5662
#> [2,] 181.56623 841.8819
#> 
#> $`98`
#>           [,1]      [,2]
#> [1,]  0.363852 -5.898595
#> [2,] -5.898595 95.625194
#> 
#> $`99`
#>           [,1]      [,2]
#> [1,]  19.24029 -39.90557
#> [2,] -39.90557  82.76669
#> 
#> $`100`
#>           [,1]      [,2]
#> [1,]   5.12201 -11.36268
#> [2,] -11.36268  25.20698
#> 
#> 
#> Slot ".ee_i":
#> $`1`
#> [1] -1.376257 -8.248583
#> 
#> $`2`
#> [1]  5.407891 19.102625
#> 
#> $`3`
#> [1] -1.921153 -6.451838
#> 
#> $`4`
#> [1] 3.3269392 0.9258588
#> 
#> $`5`
#> [1] -5.876538 24.391036
#> 
#> $`6`
#> [1] -1.645787 -7.434051
#> 
#> $`7`
#> [1] -3.1502325 -0.2187008
#> 
#> $`8`
#> [1] -1.521749 -7.826944
#> 
#> $`9`
#> [1]  4.915842 14.022842
#> 
#> $`10`
#> [1] -0.4742986 -9.9177064
#> 
#> $`11`
#> [1]  0.6458107 -9.7255941
#> 
#> $`12`
#> [1]  0.9738417 -9.1942978
#> 
#> $`13`
#> [1] -2.502699 -3.879165
#> 
#> $`14`
#> [1] -5.761424 23.051338
#> 
#> $`15`
#> [1] -1.368465 -8.269969
#> 
#> $`16`
#> [1]  0.4689809 -9.9227224
#> 
#> $`17`
#> [1] 4.027917 6.081447
#> 
#> $`18`
#> [1] -1.066858 -9.004479
#> 
#> $`19`
#> [1] -1.254727 -8.568325
#> 
#> $`20`
#> [1]  6.416199 31.024949
#> 
#> $`21`
#> [1] -3.1394168 -0.2867279
#> 
#> $`22`
#> [1]  1.651446 -7.415391
#> 
#> $`23`
#> [1] -2.380351 -4.476593
#> 
#> $`24`
#> [1]  1.615579 -7.532569
#> 
#> $`25`
#> [1] -6.22561 28.61555
#> 
#> $`26`
#> [1] -2.119561 -5.650125
#> 
#> $`27`
#> [1] -1.163730 -8.788399
#> 
#> $`28`
#> [1] 3.985266 5.739681
#> 
#> $`29`
#> [1] -1.922843 -6.445340
#> 
#> $`30`
#> [1]  1.147025 -8.827000
#> 
#> $`31`
#> [1] -1.715741 -7.198898
#> 
#> $`32`
#> [1] -3.446086  1.732846
#> 
#> $`33`
#> [1]  2.711621 -2.789775
#> 
#> $`34`
#> [1] -1.885348 -6.588128
#> 
#> $`35`
#> [1]  5.348174 18.460301
#> 
#> $`36`
#> [1]  1.727722 -7.157642
#> 
#> $`37`
#> [1] -0.6482681 -9.7224140
#> 
#> $`38`
#> [1]  1.777632 -6.982690
#> 
#> $`39`
#> [1]  -0.2051821 -10.1005658
#> 
#> $`40`
#> [1]  1.779921 -6.974547
#> 
#> $`41`
#> [1] -1.678263 -7.326097
#> 
#> $`42`
#> [1] -8.590539 63.654691
#> 
#> $`43`
#> [1]  0.5827243 -9.8030979
#> 
#> $`44`
#> [1]  2.595632 -3.405359
#> 
#> $`45`
#> [1] -3.971901  5.633332
#> 
#> $`46`
#> [1] -4.499138 10.099579
#> 
#> $`47`
#> [1] -1.793954 -6.924394
#> 
#> $`48`
#> [1] -2.109008 -5.694749
#> 
#> $`49`
#> [1]  1.631421 -7.481132
#> 
#> $`50`
#> [1]  0.4920584 -9.9005441
#> 
#> $`51`
#> [1] 4.094182 6.619665
#> 
#> $`52`
#> [1]  6.56945 33.01500
#> 
#> $`53`
#> [1]  -0.1163506 -10.1291281
#> 
#> $`54`
#> [1]  -0.1413766 -10.1226782
#> 
#> $`55`
#> [1]  0.9564243 -9.2279182
#> 
#> $`56`
#> [1] -1.393061 -8.202045
#> 
#> $`57`
#> [1] -0.4979782 -9.8946833
#> 
#> $`58`
#> [1] -0.440095 -9.948982
#> 
#> $`59`
#> [1]  1.011784 -9.118959
#> 
#> $`60`
#> [1]  0.5113714 -9.8811648
#> 
#> $`61`
#> [1] -1.013641 -9.115197
#> 
#> $`62`
#> [1]   0.1915622 -10.1059695
#> 
#> $`63`
#> [1] -0.7536508 -9.5746760
#> 
#> $`64`
#> [1]  3.1341577 -0.3197209
#> 
#> $`65`
#> [1]  -0.0176122 -10.1423553
#> 
#> $`66`
#> [1] -2.563725 -3.569982
#> 
#> $`67`
#> [1] -1.054522 -9.030648
#> 
#> $`68`
#> [1] -2.806252 -2.267615
#> 
#> $`69`
#> [1]  0.7656052 -9.5565141
#> 
#> $`70`
#> [1] 3.340962 1.019364
#> 
#> $`71`
#> [1]  2.483733 -3.973736
#> 
#> $`72`
#> [1]  0.7610908 -9.5634064
#> 
#> $`73`
#> [1] -0.4088459 -9.9755105
#> 
#> $`74`
#> [1]  1.113373 -8.903066
#> 
#> $`75`
#> [1]  -0.2633031 -10.0733370
#> 
#> $`76`
#> [1] -8.211963 57.293667
#> 
#> $`77`
#> [1]  1.388914 -8.213584
#> 
#> $`78`
#> [1] -1.534967 -7.786543
#> 
#> $`79`
#> [1]  5.028672 15.144877
#> 
#> $`80`
#> [1]  8.629838 64.331437
#> 
#> $`81`
#> [1] -5.003455 14.891894
#> 
#> $`82`
#> [1]  2.314932 -4.783754
#> 
#> $`83`
#> [1]  0.4515138 -9.9388008
#> 
#> $`84`
#> [1] -2.139394 -5.565657
#> 
#> $`85`
#> [1]  2.436349 -4.206871
#> 
#> $`86`
#> [1]  2.788323 -2.367920
#> 
#> $`87`
#> [1] -0.4173568 -9.9684789
#> 
#> $`88`
#> [1] -1.225350 -8.641183
#> 
#> $`89`
#> [1] -4.387824  9.110336
#> 
#> $`90`
#> [1] -2.543824 -3.671627
#> 
#> $`91`
#> [1] -0.9765895 -9.1889384
#> 
#> $`92`
#> [1] -1.047831 -9.044716
#> 
#> $`93`
#> [1] 3.771018 4.077911
#> 
#> $`94`
#> [1] -1.108453 -8.913997
#> 
#> $`95`
#> [1]  7.53654 46.65678
#> 
#> $`96`
#> [1] -1.764527 -7.029111
#> 
#> $`97`
#> [1]  6.257625 29.015201
#> 
#> $`98`
#> [1]  0.6032015 -9.7788135
#> 
#> $`99`
#> [1] -4.386375  9.097620
#> 
#> $`100`
#> [1]  2.263186 -5.020655
#> 
#>