Fay's degrees of freedom corrections

Fay's degrees of freedom corrections

fay_df_correction(A_i, A, B_i, B, b = 0.75, L, version)

Arguments

A_i
a list of bread matrices per group

A
the outer "bread" matrix

B_i
a list of meat matrices per group

B
the inner "meat" matrix

b
a numeric value < 1. Defaults to 0.75 as in Fay.

L
a k x p matrix where p is the dimension of theta

version
either 1 or 2, corresponding to hat(d) or tilde(d), respectively

Value

a scalar corresponding to the estimated degrees of freedom

References

Fay, M. P., & Graubard, B. I. (2001). Small-Sample adjustments for wald-type tests using sandwich estimators. Biometrics, 57(4), 1198-1206